backend/conf/requirements.txt
Package             Version
------------------- -------
asgiref             3.4.1
Django              3.2.19
django-dotenv       1.4.2
djangorestframework 3.14.0
pip                 23.1.2
psycopg2-binary     2.9.6
pypi                2.1
python-dotenv       0.20.0
pytz                2023.3
setuptools          58.0.4
sqlparse            0.4.4
typing_extensions   4.1.1
